Move PaywallPageViewInfo to own file
